[hardpy] remove axc from examples, class instrument
