+++ hardpy/__init__.py
@@ -44,6 +44,7 @@ from hardpy.pytest_hardpy.utils import (
     DuplicateFieldError,
     HTMLComponent,
     ImageComponent,
+    Instrument,
     MultistepWidget,
     NumericInputWidget,
     RadiobuttonWidget,
@@ -61,6 +62,7 @@ __all__ = [
     "DuplicateFieldError",
     "HTMLComponent",
     "ImageComponent",
+    "Instrument",
     "MultistepWidget",
     "NumericInputWidget",
     "RadiobuttonWidget",
